django/requirements.txt
beautifulsoup4==4.9.3
celery-slack==0.4.1
celery==4.4.7
django-celery-beat==2.0.0
django-celery-results==1.2.1
django-cleanup==5.0.0
django-compat==1.0.15
django-cors-headers==3.4.0
django-crispy-forms==1.9.2
django-extensions==3.0.3
django-health-check==3.12.1
django-minio-storage==0.3.8
django-oauth-toolkit==1.3.2
django-rest-framework-social-oauth2==1.1.0
Django==3.0.9
djangorestframework==3.11.0
dkpro-cassis==0.5.1
drf-yasg==1.17.1
flower==0.9.5
gunicorn==20.0.4
jsonlines==1.2.0
langdetect==1.0.8
minio==6.0.0
psutil==5.7.2
psycopg2-binary==2.8.5
pycaprio==0.2.0
pysolr==3.9.0
python-crontab==2.5.1
python-dateutil==2.8.1
python-status==1.0.1
python3-openid==3.2.0
pytz==2020.1
PyYAML==5.3.1
rdflib==5.0.0
scrapy-user-agents==0.1.1
scrapy==2.4.1
scrapyd==1.2.1
SPARQLWrapper==1.8.5
textdistance==4.2.0
tika==1.24
zipp==3.1.0
lxml==4.6.2
djangorestframework-csv==2.1.0
